[UPDATE] move hompage into component
